Rely on pick_*_point methods
Not a good idea to rely on the order of returned intersection points
of a line and a circle. Not a good idea to rely on direction either,
assuming the form will be drawn in a particular orientation. It would
be better to design algorithms and pick points closer to further away
from another reference object. For example, when getting the neck-joint
intersections with the big arcs, the points closer to the spine line
are needed.