Fix Turkish oud scale length
Form width to neck-to-bottom ratio in Turkish ouds is stated to be
3/4. This commit introduces a vertical-unit to be used for this
alteration. Points on the spine are now calculated based on this
vertical unit and the desired scale length of 58.5mm, against the
~370mm form width, is achieved.

So far, no other oud uses different vertical and horizontal units,
and it feels wrong to do this from a construction point of view.
Maybe there is another method, different to the current method
on the Lute base class, to construct this ratio of the Turkish
oud rather than resorting to changing the unit.