Remove usages of directional point pickers
As stated in c8caf6dc, relying on the orientation of the drawing to
pick a point among list of intersection points is not reliable. This
commit removes (almost) all of such usages and replacing them with
methods that rely on distance to a reference object. That way, the
desired point among multiple intersection points (i.e. when inter-
secting circles with lines, or circles) can be methodically picked.

A similar problem exists with the usage of translate_point_x and
translate_point_y methods, which makes the user dependent on the
coordinate system. However, this is a necessary evil for the time
being because traversing traditionally, through circle construction,
is proving costly, and sometimes unpredictably un-executable,
by SymPy. Maybe the cost could be alleviated through using the
simple_point method in between steps, but obviously with some
(possibly compounded) cost of accuracy.

On another note, the step that is most affected by the SymPy circle
operations' shortcomings has been the blend_two_circles method where
two circles are intersected.