Add LuteType_4 and ManolLavta_AthensMuseum class
The Manol lavtas look like type 3 or 4 lavtas. Next to the
ManolLavta_1872 in type 3, the type 4 lute is defined with
ManolLavta_AthensMuseum. Both are experiments to approach
the geometry. Also, the lavta in the museum has a larger
soundhole than other Manol lavta examples.

I'm more or less convinced that the lavtas are type 4.
However, the first blender arc seems to start above the
form center line and my current blender classes doesn't support
this. The side circle bending is tied to form center line.
Need a more generalized blending method that supports multiple
blender circles.