Initiate LavtaLowerArc
Initially I planned to have two connected tangent circles and a
blender. Now I relize that the tangent parameters need to have
some association in series so that the next circle can be centered
on a radial line through the previous circle. This radial line can
only be determined once the previous circle is defined,  with its
given radius. LowerArcBuilder and TangentParameter classes need to
accoommodate this.

Perhaps, instead of a line, reference to the previous tangent
parameter object plus a point could be given so the line can be
constructed between the previous tangent circle's center and the
given point, after the previous tangent circle is created.