Initialize a rib form builder
The ultimate goal is to create bending forms for each rib of the
lute bowl. Unless the bowl is a rotation of the soundboard outline,
each rib will have a certain bow (along the length) and twist.
If exact bending forms can be produced, then bringing such ribs
together i.e. planing the sides accordingly, should be easier.

The bow should be minimized though because it's not an easy task
to plane a bowed, especially inward, surface on a rib edge. The
bowl builder needs some refinement to analyse the bows and to
reduce or eliminate them.

The commit adds an initial version of a rib form builder. It only
plots the extended rib surface. It needs to create a 3d shape that
extends from this surface downward.