Remove end-caps from extended rib outlines
A lot was going wrong with these: misoriented lines, problems
constructing the STEP files, and so on. With this commit, the
end-caps (lines at the converging points of rib outlines at both
ends) and extensions beyond are removed.

There are some examples where the last connecting line is off,
not aligned 'well' with the line before. This must be due to some
floating point and/or geometric intersection operations. Some
interpolation is needed to treat these last connecting lines.

For the STEP file rendering, the converging points themselves are
not critical, only needed as markers perhaps. What's needed are
some extensions though, so there's some extra margin at the ends
of the physical rib while being bent. These should be implemented
through interpolation as well, ensuring good continuity. There is
enough tolerance in the physical construction to allow this.