Create new rib generation
Instead of just dividing the 'sections' in equal parts, creating
absolutely planar rib outlines, and rib surfaces without any
twist.

Also, introducing an upper and lower bound for the ribs. In the
actual physical construction, there will be an end-block on each
end of the bowl, top and bottom, that the ribs get glued onto.
No need to plot anything beyond these boundaries, which should
also avoid the inaccuracies that creep up.