Create new rib molding form script
Changed the approach of trying to build 3d molds for now. It proved
difficult to debug and fix, and it might not be even the best
approach for building the forms physically later. Started a new
method that focuses on building the parallel planes with the curve
outlines projected onto them. Thinking about building the forms by
taking these as references.