Adjust DeepBackCurve towards a half-circle
Majority, if not all, of Turkish oud bowls have a half-circle bowl
profile, except at the neck. Also slightly shallow on the bottom
side, I think. This commit adjusts the DeepBackCurve to approach
such a curve.

The exact neck profile is not a circular arc actually, but more
like flattened at the top of the curve. Next I'll try to implement
a custom curve at the neck.