Rethinking rib creation...
I'm reconsidering the initial plan of constructing wide bending
forms i.e. made up of 2 plywood panels side by side (2 x 18mm):
it won't be wide enough for a lot of cases, especially where the
bowl is flatter and ribs have a concave outline on one side.

Two panels could be wide enough for half-circle bowl profiles, but
then a single plywood panel would serve well enough as a template;
considering the bend will have some spring-back anyway.

For the case of concave-outlines ribs, for flat-ish bowls, maybe
it's better to start with a wide veneer instead of strips[1].
Depending on the thickness of the veneer, 2 or 3 plys could be
laminated in a single form, then cut to actual, consecutive* rib
slices where each slice is getting progressively more acute (or
width-wise concave). This is what's seen when plotting each rib
of a flat bowl. Maybe there is even a way to define the cutting
planes based on a bowl design to begin with. Or another approach
could be to just let the bowl profile be dictated by the relation
of consecutive ribs to each other, and to take care for the skirt
ribs to converge to a soundboard outline shape (but how...).

Some wide laminating forms would be required, out of muliple
stacked 18mm plywood panels, to create heavily concave ribs.

[1]The strips would probably have to be pretty wide to accommodate
a uniformly flat (e.g. circular arc much less than 180 degrees)
bowl anyway. I guess if the bowl is regionally flat i.e. at the
top, this could be accommodated easier, by less concave ribs.