Rethinking concave rib outlines...
- Add compute_rib_blank_width width
Continuing with the thinking from the previous commit, added a
helper method to see how wide initial blanks will have to be.

I imagine this might be useful to gauge how feasible flat bowls
ribs are, perhaps also a gateway to modifying the side-plane
picking to serve building wide rib forms: likely for veneer
lamination option, but not necessarily. The initial experiment
on lavta with FlatBackCurve gave ~70mm widest, which is not crazy.
Gotta double-check this yet though.

Edit: Been thinking on this concave rib outline issue. In short,
ribs tend to turn into width-wise (?) concave profiles when bowls
are flattened. Like a C-shape over the narrow edge, bending it
sideways. However, this is directly related to the fact that I am
terminating rib ends, at the form bottom, in a single point.
What I finally noticed by looking at some in-progress images (by
Matias Crom, @luthiermatiascrom), and thinking about skirt ribs,
is that the ribs don't converge at a single point at all. Their
'tips' land on various positions relative to the bowl edge (i.e.
middle ribs lower while side ribs higher) so the would-be concave
space is distributed between the ribs. Then these ribs are trimmed
to the outside outline of the last ribs (i.e. that the skirt ribs
will be adjacent to), or another strategy (e.g. straight line at
some height?) that creates a suitable outline to attach skirt ribs
to. By doing all this, there should be no need to have concave
ribs at all. And only after, individual rib forms would really be
useful, because Turkish ouds are just contructed with a single rib
form/profile and a semi-circular bowl profile[1].

The rib construction needs some redesign... More control over
the rib geometry is needed, maybe setting how tall/wide is a rib,
regardless of where it will terminate on the bowl, rather than
focusing on the ending point. Then some options on skirting
strategy (e.g. straight or trimmed by last rib outline).

I also would like controls to experiment with, replotting live.
For example, change the shape-parameters of a rib while other ribs
are re-positioned accordingly.

[1]Just a soundboard outline seems enough for Turkish ouds, really
(probably for Arabic ouds as well). Maybe the only remaining
difficulty is the flattening of the shape right at the neck: the
neck profile is a flatter curve. I wonder if the uplift from the
skirting, i.e. non-skirt ribs placed away from the soundboard,
naturally allows for flattening at the neck plane. Assuming that
the ribs are pre-bent: with skirting, the middle rib moves towards
the neck, tightening the curve and reducing the thickness of the
neck. The last (non-skirt) ribs move sideways, not changing their
curve function along their surface much, remaining closer to the
soundboard outline, and so keeping the neck width intact. And all
the other ribs falling on a continuum between the behaviors of the
middle rib and the last ribs.